Much improved chat
chatbox can now be docked/undocked, and messages stay forever in docked mode and disappear in undocked mode. Notifications are now stylistically different than messages. It can also be changed from left to right or right to left. use "\dock" to dock. "\float" to undock. "\dock [side]" will dock on that side and same for float.