cookies and style
separated dock / side cookies.
The floating textentry box now becomes almost invisible except when in focus.
The textentry gains focus on pressing enter. loses focus on pressing enter while empty.
The chat now defaults to docked mode so that the information about pressing \? to see list of commands is always visible (as opposed to just 15 seconds) so they can learn to float the chat whereas learning to dock would be confusing and rushed because of disappearing messages.