Added notetaking commands
Are they a good idea? idk. I know the time I was with Cal, we would have used them. They're only local... not shared between people. I think that's good tho.

Example use-case: you are watching the movie and want to keep watching but you want to come back to a specific spot later to see something in more detail, so you take a note of the current timestamp and what you wanted to look at