changed url detection regex because prev one wuld noticably slow down on non-matching words longer than 25 characters
eg. it would pause for at least 1 second on "12346789012345678901234567890"

now it only has the potential to pause on long things that start with "http(s)://" and will really only slow down if it's long and inconvenient and doesn't match. Will try to make more improvements in the future to make sure this doesn't happen at all. Maybe switch away from regex.